+++ cogs/youtube.py
@@ -1,51 +1,13 @@
 import asyncio
 import discord
-import youtube_dl
+
+import sys
 
 from discord.ext import commands, tasks
 from discord.utils import get
 
-# Suppress noise about console usage from errors
-youtube_dl.utils.bug_reports_message = lambda: ''
-
-ytdl_format_options = {
-    'format': 'bestaudio/best',
-    'outtmpl': '%(extractor)s-%(id)s-%(title)s.%(ext)s',
-    'restrictfilenames': True,
-    'noplaylist': True,
-    'nocheckcertificate': True,
-    'ignoreerrors': False,
-    'logtostderr': False,
-    'quiet': True,
-    'no_warnings': True,
-    'default_search': 'auto',
-    'source_address': '0.0.0.0'  # bind to ipv4 since ipv6 addresses cause issues sometimes
-}
-
-ytdl = youtube_dl.YoutubeDL(ytdl_format_options)
-
-
-class YTDLSource(discord.PCMVolumeTransformer):
-    def __init__(self, source, *, data, volume=0.5):
-        super().__init__(source, volume)
-
-        self.data = data
-
-        self.title = data.get('title')
-        self.url = data.get('url')
-
-    @classmethod
-    async def from_url(cls, url, ffmpeg_options, pre_op, *, loop=None, stream=True,):
-        loop = loop or asyncio.get_event_loop()
-        data = await loop.run_in_executor(None, lambda: ytdl.extract_info(url, download=not stream))
-
-        if 'entries' in data:
-            # take first item from a playlist
-            data = data['entries'][0]
-
-        filename = data['url'] if stream else ytdl.prepare_filename(data)
-        return cls(discord.FFmpegPCMAudio(filename, before_options=pre_op, **ffmpeg_options), data=data)
-
+sys.path.insert(0,"..")
+from lib.YTDLSource import YTDLSource
 
 class Music(commands.Cog):
     def __init__(self, bot):
@@ -190,9 +152,5 @@ class Music(commands.Cog):
             await ctx.send("No music playing failed to stop")
 
 
-bot = commands.Bot(command_prefix=commands.when_mentioned_or("!"),
-                   description='Relatively simple music bot example')
-
-
-def setup(client):
-    client.add_cog(Music(bot))
+async def setup(client):
+    await client.add_cog(Music(client))
